Add the Rust integration testing cargo generate template to this repo
Previously, the intention was to use the fuellabs/sway-test-rs git
repository as the source of the cargo generate template for Rust
integration testing. Instead, I've moved the template into this
repository in order to be able to test the template as a part of the
Sway CI and make sure it is always up to date.